pyproject.toml
[tool.poetry]
name = "dataverse-external-service-poc"
version = "0.1.0"
description = ""
authors = ["ekoi <eko.indarto@dans.knaw.nl>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
